+++ components/molecules/TagInputWithList/index.ts
@@ -1,0 +1,1 @@
+export { TagInputWithList } from './TagInputWithList';
